ПР-11

+++ src/utils/Api.js
@@ -19,7 +19,7 @@ class Api {
       return this._processingServerResponse(res);
     });
   }
-  addNewCard({ name, link }) {
+  addNewCard(name, link) {
     return fetch(`${this._link}cards`, {
       headers: this._headers,
       method: "POST",
@@ -43,14 +43,11 @@ class Api {
       return this._processingServerResponse(res);
     });
   }
-  sendUserData(profileData) {
+  sendUserData(userName, userAbout) {
     return fetch(`${this._link}users/me`, {
       headers: this._headers,
       method: "PATCH",
-      body: JSON.stringify({
-        name: profileData.username,
-        about: profileData.description,
-      }),
+      body: JSON.stringify({ name: userName, about: userAbout }),
     }).then((res) => {
       return this._processingServerResponse(res);
     });
@@ -64,21 +61,22 @@ class Api {
       return this._processingServerResponse(res);
     });
   }
-  putCardLike(cardId) {
-    return fetch(`${this._link}cards/${cardId}/likes`, {
-      headers: this._headers,
-      method: "PUT",
-    }).then((res) => {
-      return this._processingServerResponse(res);
-    });
-  }
-  deleteCardLike(cardId) {
-    return fetch(`${this._link}cards/${cardId}/likes`, {
-      headers: this._headers,
-      method: "DELETE",
-    }).then((res) => {
-      return this._processingServerResponse(res);
-    });
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._link}cards/${cardId}/likes`, {
+        headers: this._headers,
+        method: "PUT",
+      }).then((res) => {
+        return this._processingServerResponse(res);
+      });
+    } else {
+      return fetch(`${this._link}cards/${cardId}/likes`, {
+        headers: this._headers,
+        method: "DELETE",
+      }).then((res) => {
+        return this._processingServerResponse(res);
+      });
+    }
   }
 }
 const apiConnect = new Api(apiFindings);
